crawling/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
redis==5.1.1
google-cloud-storage==2.18.2